requirements.txt
streamlit
google-genai
groq
supabase
arxiv
requests
pandas
plotly
python-dotenv
numpy
pyarrow==14.0.2
